// frontend/src/app/core/api/model/ref-teacher-specialty-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RefTeacherSpecialtyResponse { 
    id?: number;
    code?: string;
    label?: string;
    description?: string;
    teacherCount?: number;
}