// frontend/src/app/core/api/model/ref-teacher-specialty-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RefTeacherSpecialtyRequest { 
    code: string;
    label: string;
    description?: string;
}